# openapi.json
{"openapi":"3.0.3","info":{"title":"attendance-server","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/api/v1/checkin":{"post":{"tags":["records"],"operationId":"check_in","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckInRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.dto.CheckInResponse"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}},"/api/v1/debug/dbpath":{"get":{"tags":["health"],"operationId":"db_info","responses":{"200":{"description":""}}}},"/api/v1/debug/stats":{"get":{"tags":["health"],"operationId":"stats","responses":{"200":{"description":""}}}},"/api/v1/health":{"get":{"tags":["health"],"operationId":"health","responses":{"200":{"description":""}}}},"/api/v1/locations":{"get":{"tags":["locations"],"operationId":"get_locations","responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Location"}}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}},"post":{"tags":["locations"],"operationId":"create_location","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLocationRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}},"/api/v1/locations/{id}":{"delete":{"tags":["locations"],"operationId":"delete_location","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":""},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}},"patch":{"tags":["locations"],"operationId":"update_location","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLocationRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}},"/api/v1/login":{"post":{"tags":["auth"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}},"/api/v1/records":{"get":{"tags":["records"],"operationId":"get_records","responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/crate.domain.AttendanceRecord"}}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}},"/api/v1/records/admin/{admin_id}":{"get":{"tags":["records"],"operationId":"get_records_by_admin","parameters":[{"name":"admin_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/crate.domain.AttendanceRecord"}}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}},"/api/v1/users":{"get":{"tags":["users"],"operationId":"get_users","responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}},"post":{"tags":["users"],"operationId":"create_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}},"/api/v1/users/{id}":{"delete":{"tags":["users"],"operationId":"delete_user","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":""},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}},"/api/v1/users/{id}/location":{"get":{"tags":["users"],"operationId":"get_user_location","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/crate.domain.Location"}],"nullable":true}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}},"patch":{"tags":["users"],"operationId":"update_user_location","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserLocationPayload"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/crate.error.ErrorBody"}}}}}}}},"components":{"schemas":{"AttendanceRecord":{"type":"object","required":["id","userId","locationId","latitude","longitude","timestamp","status"],"properties":{"createTime":{"type":"string","format":"date-time","nullable":true},"errorMessage":{"type":"string","nullable":true},"id":{"type":"string"},"latitude":{"type":"number","format":"double"},"locationId":{"type":"string"},"longitude":{"type":"number","format":"double"},"status":{"$ref":"#/components/schemas/AttendanceStatus"},"timestamp":{"type":"integer","format":"int64"},"updateTime":{"type":"string","format":"date-time","nullable":true},"userId":{"type":"string"}}},"AttendanceStatus":{"type":"string","enum":["success","failed"]},"CheckInRequest":{"type":"object","required":["user_id","latitude","longitude"],"properties":{"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"user_id":{"type":"string"}}},"CheckInResponse":{"type":"object","required":["success"],"properties":{"message":{"type":"string","nullable":true},"record":{"allOf":[{"$ref":"#/components/schemas/AttendanceRecord"}],"nullable":true},"success":{"type":"boolean"}}},"CreateLocationRequest":{"type":"object","required":["name","latitude","longitude","radius","adminId"],"properties":{"adminId":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"name":{"type":"string"},"radius":{"type":"number","format":"double"}}},"CreateUserRequest":{"type":"object","required":["username","password","role"],"properties":{"adminId":{"type":"string","nullable":true},"password":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"},"username":{"type":"string"}}},"ErrorBody":{"type":"object","required":["error"],"properties":{"error":{"type":"string"}}},"Location":{"type":"object","required":["id","name","latitude","longitude","radius","adminId"],"properties":{"adminId":{"type":"string"},"createTime":{"type":"string","format":"date-time","nullable":true},"id":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"name":{"type":"string"},"radius":{"type":"number","format":"double"},"updateTime":{"type":"string","format":"date-time","nullable":true}}},"LoginRequest":{"type":"object","required":["username","password"],"properties":{"password":{"type":"string"},"username":{"type":"string"}}},"LoginResponse":{"type":"object","required":["success"],"properties":{"message":{"type":"string","nullable":true},"success":{"type":"boolean"},"token":{"type":"string","nullable":true},"user":{"allOf":[{"$ref":"#/components/schemas/User"}],"nullable":true}}},"UpdateLocationRequest":{"type":"object","properties":{"latitude":{"type":"number","format":"double","nullable":true},"longitude":{"type":"number","format":"double","nullable":true},"name":{"type":"string","nullable":true},"radius":{"type":"number","format":"double","nullable":true}}},"User":{"type":"object","required":["id","username","password","role"],"properties":{"adminId":{"type":"string","nullable":true},"createTime":{"type":"string","format":"date-time","nullable":true},"id":{"type":"string"},"locationId":{"type":"string","nullable":true},"password":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"},"updateTime":{"type":"string","format":"date-time","nullable":true},"username":{"type":"string"}}},"UserRole":{"type":"string","enum":["admin","user"]}}},"tags":[{"name":"auth","description":"authentication"},{"name":"users","description":"user management"},{"name":"locations","description":"location management"},{"name":"records","description":"attendance records"},{"name":"health","description":"health endpoints"}]}